+++ src/readers_xrd/__init__.py
@@ -1,0 +1,1 @@
+from .readers import *
